chore(metrics): add kafka topic in notification sent (#72)
* chore(metrics): add kafka topic in notif sent

To create composite alarms, we must ensure that we can aggregate
by the same info, which in this case is kafka topic. The idea is
that we have alarms if kafka lag is decreasing (if Pusher is
consuming) while we are sending this messages to the provider.
IOW, if we are fanning-in <> fanning-out accordingly, regardless
of the pace

* chore(go): update to 1.23

* chore(docker): setup to work with ARM

* test: fix flaky tests